+++ sickrage/providers/torrent/iptorrents.py
@@ -19,12 +19,13 @@
 from __future__ import unicode_literals
 
 import re
+from urlparse import urljoin
 
 from requests.utils import dict_from_cookiejar
 
 import sickrage
 from sickrage.core.caches.tv_cache import TVCache
-from sickrage.core.helpers import bs4_parser, convert_size, try_int
+from sickrage.core.helpers import bs4_parser, convert_size, try_int, validate_url
 from sickrage.providers import TorrentProvider
 
 
@@ -48,6 +49,8 @@ class IPTorrentsProvider(TorrentProvider):
 
         self.categories = '73=&60='
 
+        self.custom_url = ""
+
         self.cache = TVCache(self, min_time=10)
 
     def login(self):
@@ -64,7 +67,15 @@ class IPTorrentsProvider(TorrentProvider):
 
         login_params = {'username': self.username, 'password': self.password, 'login': 'submit'}
 
-        response = sickrage.srCore.srWebSession.post(self.urls['login'], data=login_params, timeout=30)
+        login_url = self.urls['login']
+        if self.custom_url:
+            if not validate_url(self.custom_url):
+                sickrage.srCore.srLogger.warning("Invalid custom url: {}".format(self.custom_url))
+                return False
+
+            login_url = urljoin(self.custom_url, self.urls['login'].split(self.urls['base_url'])[1])
+
+        response = sickrage.srCore.srWebSession.post(login_url, data=login_params, timeout=30)
         if not response.ok:
             sickrage.srCore.srLogger.warning("[{}]: Unable to connect to provider".format(self.name))
             return False
@@ -102,12 +113,19 @@ class IPTorrentsProvider(TorrentProvider):
                     sickrage.srCore.srLogger.debug("Search string: %s " % search_string)
 
                 # URL with 50 tv-show results, or max 150 if adjusted in IPTorrents profile
-                searchURL = self.urls['search'] % (self.categories, freeleech, search_string)
-                searchURL += ';o=seeders' if mode != 'RSS' else ''
-                sickrage.srCore.srLogger.debug("Search URL: %s" % searchURL)
+                search_url = self.urls['search'] % (self.categories, freeleech, search_string)
+                search_url += ';o=seeders' if mode != 'RSS' else ''
+
+                if self.custom_url:
+                    if not validate_url(self.custom_url):
+                        sickrage.srCore.srLogger.warning("Invalid custom url: {}".format(self.custom_url))
+                        return results
+                    search_url = urljoin(self.custom_url, search_url.split(self.urls['base_url'])[1])
+
+                sickrage.srCore.srLogger.debug("Search URL: %s" % search_url)
 
                 try:
-                    data = sickrage.srCore.srWebSession.get(searchURL).text
+                    data = sickrage.srCore.srWebSession.get(search_url).text
                 except Exception:
                     sickrage.srCore.srLogger.debug("No data returned from provider")
                     continue